requirements.txt
urwid==2.0.1
PyYAML==3.12
pylint==1.8.4
